Move the primary implementations to `main` crate to organize codes.
fastly CLI (at least v1.7.1) does not support a way to work with Cargo Workspace.
This mean that we need to keep a crate as an entry point on the directory
which we run `fastly compute build`, `deploy`, etc. because
fastly CLI check & use `package.name` field in Cargo.toml in the current working directory
to indentify the build artifact (wasm file) which fastly CLI manipulate before deploying it.