Remove `default` feature from Cargo.toml
In the prev commit, we don't use default features in dependencies by default.
But it's not clear for rust-analyzer or other toolchains which are not
integrated to our build system.

To make it simplify, this change removes all `default` feature set from
Cargo.toml in this repository.

This express that this project requires an additional way to enable
toolchain supports which is not added to our build system.

Drawbacks
--------

1. rust-analyzer does not support editor agnostic setting.
   A developer that do not use vscode would requires additional config.
   rust-lang/rust-analyzer#11010
2. This requires to commit `.vscode/settings.json`.