ci: Drop the step to call fastly compute validate
This command's target artifact is built upon `fastly compute pack` in the just previous step.
In this situation, I think we can trust the artifact
as a valid format for fastly compute service and we can remove the step from normal CI.

To detect an error, we place this command to the step in the deployment workflow.