Lock viceroy_version more precisely for fastly compute serve
By the CLI source code, fastly CLI always attempt to check the update of Viceory
even if there is a latest Viceroy.
https://github.com/fastly/cli/blob/8d58ba09aaff2af9d4875b8bafe6b7b651694b6d/pkg/commands/compute/serve.go#L445-L524

This sometime causes a flaky behavior for CI.
And also it requires that local
machine always connect to a network.

To make more reliable, we specify a viceroy_version for integration test.