Save cache of dependencies for CI only if on _trunk_ branch
We do this to aim to improve a cache hit.

For example, dependabot's pull requests sometimes shoot down a cache hit ratio
by changing a lock file with multiple variants.
hen we cannot reuse cache ideally because the lockfile is changed high frequently under such situation.
Cache utilization is down.

So we give up to save the cache on each of pull requests.